Moved the functions to its own library rs Moved the macro to its own rs Added a tests directory and a simple unit test file for now
Jira-Id: COR-1672
