+++ state-compare/Cargo.toml
@@ -19,3 +19,8 @@ tokio = {version = "1.20", features = ["rt-multi-thread"]}
 tonic = { version = "0.12", features = ["tls-roots", "tls"] }
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+
+# need these to make the tests printing the logs from the token_compare.rs
+# run cargo test -- --nocapture
+log = "0.4"
+env_logger = "0.11"
